Simplify the remote translations client namespace
Since Microsoft is no longer used, the Llm:: namespace under RemoteTranslations
is no longer needed to separate providers, so RemoteTranslations::Llm::Client
does not add any useful information.
