+++ spec/lib/remote_translations/client_spec.rb
@@ -1,10 +1,10 @@
 require "rails_helper"
 
-describe RemoteTranslations::Llm::Client do
+describe RemoteTranslations::Client do
   let(:chat) { instance_double(RubyLLM::Chat, ask: double(content: "translated")) }
   let(:config)  { instance_double(RubyLLM::Configuration) }
   let(:context) { instance_double(RubyLLM::Context, chat: chat, config: config) }
-  let(:client) { RemoteTranslations::Llm::Client.new }
+  let(:client) { RemoteTranslations::Client.new }
 
   before do
     Setting["llm.provider"] = "OpenAI"
